+++ src/gort/devices/spec.py
@@ -246,7 +246,12 @@ class Spectrograph(GortDevice):
 
         status = await self.status(simple=True)
         names = status["status_names"]
-        return "IDLE" in names and "READOUT_PENDING" not in names
+
+        for status in ["READOUT_PENDING", "ERROR"]:
+            if status in names:
+                return False
+
+        return "IDLE" in names
 
     async def is_exposing(self):
         """Returns :obj:`True` if the spectrograph is exposing."""