+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Next version
+
+### 🔧 Fixed
+
+* Fixed a bug that would prevent a new SJD to trigger an update of the ephemeris and calibrations.
+
 ## 1.1.2 - November 11, 2024
 
 ### ✨ Improved