+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Run some pre-observing checks before calling each `GortObserver.observe_tile()` in the `ObserverOverwatcher`. Currently only checks if the spectrographs have an error state and resets them.
 * Handle `SPECTROGRAPH_NOT_IDLE` errors in the troubleshooter.
 * Disable the Overwatcher and cancel observations if the dome fails to move.
+* Add retries for safe enclosure operations.
 
 ### 🔧 Fixed